Refactor QR Code page and components: Split QR code functionality into a new QrCodeClient component for better organization. Updated InteractiveMap to use useState for pan/zoom instead of useSpring, and improved mobile handling. Adjusted TopBar styles for better UI consistency.
