+++ frontend/student-web/app/(app)/community/coin-hunting/_components/InteractiveMap.tsx
@@ -2,7 +2,7 @@
 
 import React, { useRef, useState, useEffect, useCallback } from 'react';
 import { useGesture } from '@use-gesture/react';
-import { animated, useSpring } from '@react-spring/web';
+import { animated } from '@react-spring/web';
 
 type InteractiveMapProps = {
   onImageLoad?: (imageSize: { width: number; height: number }) => void;
@@ -15,21 +15,19 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
   const [imageSize, setImageSize] = useState({ width: 6000, height: 2469 });
   const [containerSize, setContainerSize] = useState({ width: 0, height: 0 });
 
-  // เพิ่ม pan (x, y) และ scale
-  const [{ x, y, scale }, api] = useSpring(() => ({
-    x: 0,
-    y: 0,
-    scale: 1,
-    config: { tension: 300, friction: 30 },
-  }));
+  // ใช้ useState แทน useSpring สำหรับ pan/zoom
+  const [pan, setPan] = useState({ x: 0, y: 0 });
+  const [scale] = useState(1); // ล็อก scale ไว้ที่ 1
 
-  // Load image size once
+  // Load image size once, and set loaded flag
+  const [isImageLoaded, setIsImageLoaded] = useState(false);
   useEffect(() => {
     const img = new window.Image();
-    img.src = '/images/map.png'; // Make sure it's in /public/images/
+    img.src = '/images/map.png';
     img.onload = () => {
       setImageSize({ width: img.width, height: img.height });
       onImageLoad?.({ width: img.width, height: img.height });
+      setIsImageLoaded(true);
     };
   }, []);
 
@@ -49,26 +47,18 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
   // Detect mobile device
   const isMobile = typeof window !== 'undefined' && /Mobi|Android|iPhone|iPad|iPod|Mobile/i.test(navigator.userAgent);
 
-  // Center map on load (reset x, y)
-  useEffect(() => {
-    if (!containerRef.current || !imageSize.width || !imageSize.height) return;
-    const cW = window.innerWidth;
-    const cH = window.innerHeight;
-    const iW = imageSize.width;
-    const iH = imageSize.height;
-    // Center to (0,0) for now, or calculate initial x, y if needed
-    api.start({ x: 0, y: 0 });
-  }, [imageSize.width, imageSize.height, api]);
+  // กำหนด initial scale ให้แผนที่แสดงผลแบบ zoom-out
+  const initialScale = 0.7; // ปรับค่านี้เพื่อความกว้างที่ต้องการ
 
   // Helper: clamp value
   const clamp = (value: number, min: number, max: number) => Math.max(min, Math.min(value, max));
 
   // ปรับ clamp pan/zoom ไม่ให้ออกนอกขอบภาพ (แก้ getBounds)
-  const getBounds = (scaleValue: number) => {
+  const getBounds = () => {
     const cW = window.innerWidth;
     const cH = window.innerHeight;
-    const scaledWidth = imageSize.width * scaleValue;
-    const scaledHeight = imageSize.height * scaleValue;
+    const scaledWidth = imageSize.width * initialScale;
+    const scaledHeight = imageSize.height * initialScale;
     let minX = 0, maxX = 0, minY = 0, maxY = 0;
     if (scaledWidth > cW) {
       minX = cW - scaledWidth;
@@ -89,40 +79,21 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
   const bind = useGesture(
     {
       onDrag: ({ offset: [dx, dy], last, memo }) => {
-        const s = scale.get();
-        // ปรับ dx, dy ให้สัมพันธ์กับ scale
-        const adjDx = dx / s;
-        const adjDy = dy / s;
-        const bounds = getBounds(s);
+        // scale = 1 เสมอ แต่แสดงผลด้วย initialScale
+        const adjDx = dx;
+        const adjDy = dy;
+        const bounds = getBounds();
         const clampedX = clamp(adjDx, bounds.minX, bounds.maxX);
         const clampedY = clamp(adjDy, bounds.minY, bounds.maxY);
-        api.start({ x: clampedX, y: clampedY });
-        return memo;
-      },
-      onPinch: ({ offset: [d], origin, memo, event }) => {
-        // d: scale value
-        const scaleValue = Math.max(1, Math.min(d, 3));
-        // Clamp x, y to new bounds after zoom
-        const bounds = getBounds(scaleValue);
-        // Get current x, y
-        const currentX = x.get();
-        const currentY = y.get();
-        const clampedX = clamp(currentX, bounds.minX, bounds.maxX);
-        const clampedY = clamp(currentY, bounds.minY, bounds.maxY);
-        api.start({ scale: scaleValue, x: clampedX, y: clampedY });
+        setPan({ x: clampedX, y: clampedY });
         return memo;
       },
     },
     {
       drag: {
-        from: () => [x.get(), y.get()],
+        from: () => [pan.x, pan.y],
         filterTaps: true,
         ...(isMobile ? { pointer: { touch: true } } : {}),
-      },
-      pinch: {
-        from: () => [scale.get(), 0],
-        scaleBounds: { min: 1, max: 3 },
-        rubberband: false,
       },
     }
   );
@@ -143,9 +114,9 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
       ref={containerRef}
       {...bind()}
       className="fixed top-0 left-0 w-screen h-screen bg-black overflow-hidden touch-auto z-0"
-      style={{}}
+      style={{ touchAction: 'none' }}
     >
-      <animated.div
+      <div
         style={{
           width: imageSize.width,
           height: imageSize.height,
@@ -153,9 +124,7 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
           left: 0,
           top: 0,
           transformOrigin: 'top left',
-          scale,
-          x,
-          y,
+          transform: `translate(${pan.x}px, ${pan.y}px) scale(${initialScale})`,
         }}
         className=""
       >
@@ -170,7 +139,7 @@ export default function InteractiveMap({ onImageLoad, children }: InteractiveMap
         <div className="absolute inset-0 z-20">
           {children}
         </div>
-      </animated.div>
+      </div>
     </div>
   );
 }
