+++ frontend/student-web/app/(app)/community/coin-hunting/_components/TopBar.tsx
@@ -47,9 +47,9 @@ const styles: { [key: string]: React.CSSProperties } = {
     paddingRight: 20,
   },
   iconBtn: {
-    width: 36,
-    height: 36,
-    borderRadius: 18,
+    width: 46,
+    height: 46,
+    borderRadius: 24,
     backgroundColor: 'rgba(0,0,0,0.15)',
     display: 'flex',
     justifyContent: 'center',
@@ -59,8 +59,8 @@ const styles: { [key: string]: React.CSSProperties } = {
     cursor: 'pointer',
   },
   stampIcon: {
-    width: 24,
-    height: 24,
+    width: 36,
+    height: 36,
     objectFit: 'contain',
   },
 };
