+++ frontend/student-web/app/(app)/qrcode/page.tsx
@@ -1,292 +1,10 @@
-'use client';
+import { Suspense } from "react";
+import QrCodeClient from "./QrCodeClient";
 
-import { QRCodeCanvas } from 'qrcode.react';
-import { Button, addToast } from '@heroui/react';
-import { Save } from 'lucide-react';
-import { useRef, useState } from 'react';
-import { useEffect } from 'react';
-import { useSearchParams } from 'next/navigation';
-
-import QRCodeSkeleton from './_components/QRCodeSkeleton';
-
-import { useProfile } from '@/hooks/useProfile';
-
-// เพิ่ม import สำหรับ QR Scanner
-import dynamic from 'next/dynamic';
-import { apiRequest } from '@/utils/api';
-import { useRouter } from 'next/navigation';
-// import { usePopupDialog } from '@/components/ui/dialog'; // ลบออก
-import { CustomFinderTracker } from './_components/CustomFinderTracker';
-
-// ใช้ dynamic import เพื่อหลีกเลี่ยง SSR error
-const QrScanner = dynamic(
-  () => import('@yudiel/react-qr-scanner').then(mod => mod.Scanner),
-  { ssr: false }
-);
-
-export default function QRCodePage() {
-  const { user, loading } = useProfile();
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const [tab, setTab] = useState<'show' | 'scan'>('show');
-  const [scanResult, setScanResult] = useState<string | null>(null);
-  const [scanning, setScanning] = useState(false);
-  const scanningRef = useRef(false);
-  const router = useRouter();
-  // const { openDialog } = usePopupDialog(); // ลบออก
-  const searchParams = useSearchParams();
-
-  useEffect(() => {
-    const tabParam = searchParams.get('tab');
-    if (tabParam === 'scan' || tabParam === 'show') {
-      setTab(tabParam);
-      resetScanner();
-    }
-    // eslint-disable-next-line
-  }, [searchParams]);
-
-  // --- Utility Functions ---
-  function showErrorAlert(message: string) {
-    // ถ้า error มีคำว่า internal server ให้แสดงข้อความใหม่
-    const displayMessage =
-      message && message.toLowerCase().includes('internal server')
-        ? 'เกิดข้อผิดพลาดระหว่างการแสกน'
-        : message;
-    addToast({
-      title: displayMessage,
-      color: 'danger',
-    });
-    resetScanner();
-  }
-
-  function validateAndParseHLLCQR(data: string): { qrPath: string, qrPayload: string } | null {
-    if (!data.startsWith('hllc:')) return null;
-    const jsonStr = data.slice(5);
-    let qrObj: Record<string, string>;
-    try {
-      qrObj = JSON.parse(jsonStr);
-    } catch {
-      return null;
-    }
-    const [qrPath, qrPayload] = Object.entries(qrObj)[0] || [];
-    if (!qrPath || !qrPayload) return null;
-    return { qrPath, qrPayload };
-  }
-
-  function resetScanner() {
-    setScanning(false);
-    scanningRef.current = false;
-    setScanResult(null);
-  }
-
-  // --- Main scan handler ---
-  const handleBarcodeScanned = async (data: string) => {
-    if (scanningRef.current) return;
-    scanningRef.current = true;
-    setScanning(true);
-
-    // validate QR
-    const parsed = validateAndParseHLLCQR(data);
-    if (!parsed) {
-      showErrorAlert('QR นี้ไม่ใช่ของ HLLC หรือรูปแบบไม่ถูกต้อง');
-      return;
-    }
-    const { qrPath, qrPayload } = parsed;
-
-    try {
-      if (!user?._id && !user?.username) {
-        showErrorAlert('ไม่พบข้อมูลผู้ใช้');
-        return;
-      }
-      // Get location from browser
-      const getLocation = (): Promise<{ latitude: number; longitude: number }> => {
-        return new Promise((resolve, reject) => {
-          if (!navigator.geolocation) {
-            reject(new Error('เบราว์เซอร์ไม่รองรับการขอตำแหน่งที่ตั้ง'));
-            return;
-          }
-          navigator.geolocation.getCurrentPosition(
-            pos => resolve({ latitude: pos.coords.latitude, longitude: pos.coords.longitude }),
-            err => reject(new Error('กรุณาอนุญาตให้เข้าถึงตำแหน่งเพื่อเช็คอิน')),
-            { enableHighAccuracy: true, timeout: 10000 }
-          );
-        });
-      };
-      let userLat = 0, userLong = 0;
-      try {
-        const loc = await getLocation();
-        userLat = loc.latitude;
-        userLong = loc.longitude;
-      } catch (e: any) {
-        showErrorAlert(e.message || 'ไม่สามารถดึงตำแหน่งที่ตั้งได้');
-        return;
-      }
-      const userId = user._id || user.username;
-      const path = `/${qrPath}`;
-      let body: any = { user: userId, userLat, userLong };
-      const payloadTrimmed = qrPayload.trim();
-      if (payloadTrimmed.startsWith('{') || payloadTrimmed.startsWith('[')) {
-        try {
-          const parsedPayload = JSON.parse(payloadTrimmed);
-          body = { ...body, ...parsedPayload };
-        } catch (e) {
-          showErrorAlert('QR code payload ไม่ถูกต้อง');
-          return;
-        }
-      } else {
-        body = qrPath === 'coin-collections/collect'
-          ? { ...body, landmark: qrPayload.trim() }
-          : { ...body, id: qrPayload.trim() };
-      }
-      const res = await apiRequest<any>(path, 'POST', body);
-      if (qrPath === 'coin-collections/collect') {
-        if (res.statusCode === 200 || res.statusCode === 201) {
-          router.push(`/community/coin-hunting?modal=success${res?.data?.evoucher?.code ? `&code=${encodeURIComponent(res.data.evoucher.code)}` : ''}`);
-        } else if (res.statusCode === 409 || (res.message && res.message.toLowerCase().includes('already'))) {
-          router.push('/community/coin-hunting?modal=alert&type=already-collected');
-        } else if (res.statusCode === 204 || (res.message && res.message.toLowerCase().includes('no new evoucher'))) {
-          router.push('/community/coin-hunting?modal=alert&type=no-evoucher');
-        } else if (res.statusCode === 403 || (res.message && res.message.toLowerCase().includes('too far'))) {
-          router.push('/community/coin-hunting?modal=alert&type=too-far');
-        } else if (res.statusCode === 429 || (res.message && res.message.toLowerCase().includes('cooldown'))) {
-          router.push('/community/coin-hunting?modal=alert&type=cooldown');
-        } else {
-          showErrorAlert(res.message || 'Check in failed');
-        }
-        resetScanner();
-        return;
-      }
-      // กรณีอื่นๆ
-      if (res.statusCode !== 200 && res.statusCode !== 201) {
-        // ดัก error ที่มี internal server
-        showErrorAlert(res.message || 'Check in failed');
-        return;
-      }
-      addToast({
-        title: 'สำเร็จ',
-        color: 'success',
-      });
-    } catch (e: any) {
-      // ดัก error ที่มี internal server
-      const errMsg = e.message && e.message.toLowerCase().includes('internal server')
-        ? 'เกิดข้อผิดพลาดระหว่างการแสกน'
-        : 'Scan failed: ' + (e.message || 'Unknown error');
-      showErrorAlert(errMsg);
-      return;
-    }
-    resetScanner();
-  };
-
-  const handleDownload = () => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const url = canvas.toDataURL('image/png');
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = `${user?.username ?? 'qrcode'}.png`;
-    link.click();
-  };
-
-  if (loading || !user) return <QRCodeSkeleton />;
-
+export default function Page() {
   return (
-    <div className="flex h-full w-full items-center justify-center">
-      <div className="flex flex-col items-center justify-center gap-6 mx-4 w-full max-w-md">
-        {/* Tab UI - pill style with blur and animated indicator */}
-        <div className="relative w-full flex items-center justify-center mb-8 select-none border border-white/30 rounded-full shadow-lg backdrop-blur-md bg-white/10" style={{height: 56}}>
-          {/* Focus pill indicator */}
-          <div
-            className={`absolute top-0 left-0 h-14 w-1/2 transition-transform duration-200 ease-out z-0 rounded-full backdrop-blur-md bg-white/20 border border-white/30 shadow-lg`}
-            style={{
-              transform: tab === 'show' ? 'translateX(0)' : 'translateX(100%)',
-              transition: 'transform 0.22s cubic-bezier(.4,1.2,.4,1)',
-            }}
-          />
-          {/* Tab buttons */}
-          <button
-            className={`relative flex-1 h-14 z-10 rounded-l-full font-bold transition-colors duration-200 text-lg
-              ${tab === 'show' ? 'text-white' : 'text-white/70'}`}
-            onClick={() => { setTab('show'); resetScanner(); }}
-            style={{ WebkitTapHighlightColor: 'transparent' }}
-          >
-            <span className="drop-shadow-sm">แสดง QR Code</span>
-          </button>
-          <button
-            className={`relative flex-1 h-14 z-10 rounded-r-full font-bold transition-colors duration-200 text-lg
-              ${tab === 'scan' ? 'text-white' : 'text-white/70'}`}
-            onClick={() => { setTab('scan'); resetScanner(); }}
-            style={{ WebkitTapHighlightColor: 'transparent' }}
-          >
-            <span className="drop-shadow-sm">สแกน QR Code</span>
-          </button>
-        </div>
-
-        {/* Content wrapper: fix size and center content, like mobile */}
-        <div className="min-h-[420px] h-[420px] w-full flex flex-col justify-center items-center">
-          {tab === 'show' ? (
-            <div
-              className="flex flex-col items-center justify-center py-8 px-4 rounded-2xl shadow-lg bg-black/20 backdrop-blur-md border border-white/30 text-center w-full h-full max-w-xl"
-            >
-              <div className="mb-4 gap-0">
-                <h1 className="text-2xl font-bold text-white">
-                  {user?.name.first + ' ' + user?.name.middle + user?.name.last || 'Your Name'}
-                </h1>
-                <p className="font-semibold text-white/80">
-                  Student ID: {user?.username || '680000000000'}
-                </p>
-              </div>
-              <QRCodeCanvas
-                ref={canvasRef}
-                bgColor="transparent"
-                fgColor="#ffffff"
-                size={200}
-                value={user?.username ?? ''}
-              />
-              <div className="mt-4 text-center px-4">
-                <p className="text-sm font-bold text-white">
-                  Please show your QR Code before joining the activity.
-                </p>
-                <Button
-                  className="bg-white text-black font-bold px-4 py-2 mt-4"
-                  radius="full"
-                  onPress={handleDownload}
-                >
-                  <Save className="mr-2" />
-                  Download QR Code
-                </Button>
-              </div>
-            </div>
-          ) : (
-            <div className="w-full max-w-xl flex flex-col items-center justify-center gap-4 bg-black/20 rounded-2xl p-4 border border-white/30 min-h-[400px] shadow-lg backdrop-blur-md">
-              <div className="w-full flex flex-col items-center justify-center">
-                <QrScanner
-                  onScan={(codes: any[]) => {
-                    if (!codes || codes.length === 0) {
-                      setScanResult('ไม่พบ QR หรือกล้องมีปัญหา');
-                      return;
-                    }
-                    const result = codes[0]?.rawValue || codes[0]?.value || '';
-                    if (result) {
-                      handleBarcodeScanned(result);
-                    } else {
-                      setScanResult('ไม่พบ QR หรือกล้องมีปัญหา');
-                    }
-                  }}
-                  constraints={{ facingMode: 'environment' }}
-                  components={{ finder: false, tracker: CustomFinderTracker, zoom: true }}
-                  styles={{ container: { width: '100%', height: '15rem', borderRadius: '0.75rem', overflow: 'hidden' } }}
-                />
-              </div>
-              <div className="mt-2 text-center min-h-[2rem]">
-                {scanResult ? (
-                  <span className="text-white font-bold">ผลลัพธ์: {scanResult}</span>
-                ) : (
-                  <span className="text-white/60">กรุณาสแกน QR Code</span>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
+    <Suspense>
+      <QrCodeClient />
+    </Suspense>
   );
 }
